[GLUTEN-10134][VL] Move GlutenExpressionOffloadTracker to common and fix ansiTest
Move GlutenExpressionOffloadTracker from spark40/spark41 shims to
gluten-ut/common, eliminating code duplication. Change ansiTest from
lazy val to protected def to avoid stale config reads when withSQLConf
changes enableAnsiFallback between tests.

Spark41 suites additionally mixin shim.GlutenTestsTrait so that
resolveExpression picks up the replace() override via Scala linearization.
