+++ gluten-ut/common/src/test/scala/org/apache/spark/sql/GlutenExpressionOffloadTracker.scala
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.shim
+package org.apache.spark.sql
 
 import org.apache.gluten.execution.ProjectExecTransformer
 
-import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Expression
