+++ gluten-ut/spark41/src/test/scala/org/apache/spark/sql/catalyst/expressions/GlutenMathExpressionsSuite.scala
@@ -16,11 +16,15 @@
  */
 package org.apache.spark.sql.catalyst.expressions
 
+import org.apache.spark.sql.GlutenExpressionOffloadTracker
 import org.apache.spark.sql.catalyst.dsl.expressions._
-import org.apache.spark.sql.shim.GlutenExpressionOffloadTracker
+import org.apache.spark.sql.shim.GlutenTestsTrait
 import org.apache.spark.sql.types._
 
-class GlutenMathExpressionsSuite extends MathExpressionsSuite with GlutenExpressionOffloadTracker {
+class GlutenMathExpressionsSuite
+  extends MathExpressionsSuite
+  with GlutenExpressionOffloadTracker
+  with GlutenTestsTrait {
   testGluten("round/bround/floor/ceil") {
     val scales = -6 to 6
     val doublePi: Double = math.Pi
