+++ gluten-ut/spark41/src/test/scala/org/apache/spark/sql/catalyst/expressions/GlutenTryCastSuite.scala
@@ -17,11 +17,12 @@
 package org.apache.spark.sql.catalyst.expressions
 
 import org.apache.spark.SparkThrowable
+import org.apache.spark.sql.GlutenExpressionOffloadTracker
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.util.DateTimeTestUtils.{withDefaultTimeZone, ALL_TIMEZONES, UTC, UTC_OPT}
 import org.apache.spark.sql.catalyst.util.DateTimeUtils.{fromJavaTimestamp, millisToMicros, TimeZoneUTC}
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.shim.GlutenExpressionOffloadTracker
+import org.apache.spark.sql.shim.GlutenTestsTrait
 import org.apache.spark.sql.types._
 import org.apache.spark.util.DebuggableThreadUtils
 
@@ -30,7 +31,10 @@ import java.util.{Calendar, TimeZone}
 
 import scala.reflect.ClassTag
 
-class GlutenTryCastSuite extends TryCastSuite with GlutenExpressionOffloadTracker {
+class GlutenTryCastSuite
+  extends TryCastSuite
+  with GlutenExpressionOffloadTracker
+  with GlutenTestsTrait {
   override protected def panoramaMeta(expression: Expression): String = expression match {
     case c: Cast => s"fromType=${c.child.dataType.simpleString},toType=${c.dataType.simpleString}"
     case _ => ""
